Cache reporter when creating ReportRequest
