+++ src/platform/browser/send-spans-fn.ts
