+++ src/platform/node/send-spans-fn.ts
