+++ test/common/create-report-request.test.ts
@@ -1,5 +1,5 @@
 import * as assert from 'assert';
-import { createReportRequest } from '../../src/create-report-request';
+import { createReportRequestFn } from '../../src/create-report-request-fn';
 import * as ls from '../../src/types';
 import { spanWithParent, reportJSON } from '../helper';
 
@@ -8,9 +8,11 @@ describe('createReportRequest', () => {
     const guid = '1234abcd1234abcd';
     const token = 'dkasjdalsjdlaksjdkaskldj';
     const tags = { foo: 'bar' };
-    const result: ls.ReportRequest = createReportRequest(guid, token, tags, [
-      spanWithParent,
-    ]);
+    const result: ls.ReportRequest = createReportRequestFn(
+      guid,
+      token,
+      tags
+    )([spanWithParent]);
 
     assert.deepEqual(result, reportJSON, 'wrong report');
   });
